Move EvakaSessionUser to a better place
It's a universal thing, not related to "SAML routes"
