+++ apigw/src/shared/express.ts
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import type express from 'express'
-import type { EvakaSessionUser } from './routes/auth/saml/types'
 import { BaseError } from 'make-error-cause'
+import { EvakaSessionUser } from './auth'
 
 export interface LogoutToken {
   // milliseconds value of a Date. Not an actual Date because it will be JSONified
